Add superclass for file choosers
